(DiamondLightSource/python-artemis#798) Fix reading OAVParameters
