+++ src/dodal/devices/oav/oav_parameters.py
@@ -10,8 +10,8 @@ from dodal.devices.oav.oav_errors import (
 from dodal.log import LOGGER
 
 OAV_CONFIG_FILE_DEFAULTS = {
-    "zoom_params_file": "/dls_sw/i03/software/gda_versions/gda_9_27/workspace_git/gda-mx.git/configurations/i03-config/xml/jCameraManZoomLevels.xml",
-    "oav_config_json": "/dls_sw/i03/software/gda_versions/gda_9_27/workspace_git/gda-mx.git/configurations/i03-config/etc/OAVCentring.json",
+    "zoom_params_file": "/dls_sw/i03/software/gda/configurations/i03-config/xml/jCameraManZoomLevels.xml",
+    "oav_config_json": "/dls_sw/i03/software/gda/configurations/i03-config/etc/OAVCentring.json",
     "display_config": "/dls_sw/i03/software/gda_versions/var/display.configuration",
 }
 
@@ -53,7 +53,7 @@ class OAVParameters:
 
         self.global_params, self.context_dicts = self.load_json(self.oav_config_json)
         self.active_params = ChainMap(
-            {}, self.global_params, self.context_dicts[self.context]
+            self.context_dicts[self.context], self.global_params
         )
         self.update_self_from_current_context()
         self.load_microns_per_pixel()
