+++ tests/devices/unit_tests/test_OAVCentring.json
@@ -29,7 +29,8 @@
 		"CannyEdgeLowerThreshold": 5.0,
 		"brightness": 20,
 		"filename": "/dls_sw/prod/R3.14.12.7/support/adPython/2-1-11/adPythonApp/scripts/adPythonMxSampleDetect.py",
-		"max_tip_distance": 300
+		"max_tip_distance": 300,
+		"minheight": 10
 	},
 	"xrayCentring": {
 		"zoom": 7.5,
